+++ source/Plugin.cs
@@ -25,6 +25,7 @@ namespace LobbyImprovements
         internal static ConfigEntry<bool> singleplayerLobbyMenu;
         
         internal static ConfigEntry<bool> saveDeleteEnabled;
+        internal static ConfigEntry<bool> savePublicEnabled;
         
         internal static ConfigEntry<bool> testerOverlayEnabled;
         internal static ConfigEntry<bool> testerOverlayModule;
@@ -83,6 +84,15 @@ namespace LobbyImprovements
             }
             
             saveDeleteEnabled = StaticConfig.Bind("Saves", "Deletion", true, "Should saves be automatically deleted when everyone dies?");
+            savePublicEnabled = StaticConfig.Bind("Saves", "Public Lobbies", false, "Should public lobbies have save files?");
+            try
+            {
+                harmony.PatchAll(typeof(PublicLobbySaves));
+            }
+            catch (Exception e)
+            {
+                StaticLogger.LogError("PublicLobbySaves Patch Failed: " + e);
+            }
             
             try
             {